# services/core-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  dtos.AuthLoginDto:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  dtos.AuthLoginSchema:
    type: object
  dtos.CreateBlogDto:
    properties:
      content:
        type: string
      name:
        type: string
      shortDescription:
        type: string
    type: object
  models.Blog:
    properties:
      content:
        type: string
      createTime:
        type: string
      id:
        type: integer
      name:
        type: string
      shortDescription:
        type: string
      updateTime:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Core Service API
  title: Core Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with credentials
      parameters:
      - description: User data
        in: body
        name: AuthLogin
        required: true
        schema:
          $ref: '#/definitions/dtos.AuthLoginDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.AuthLoginSchema'
      summary: Login
      tags:
      - auth
  /blog:
    get:
      consumes:
      - application/json
      description: Get all blog posts from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Blog'
            type: array
      summary: Get all blog posts
      tags:
      - blog
    post:
      consumes:
      - application/json
      description: Creates a new blog post
      parameters:
      - description: Blog post data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateBlogDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Blog'
      summary: Create blog post
      tags:
      - blog
swagger: "2.0"